+++ HardLight/src/HL_CreateVehicle.cpp
@@ -8,25 +8,25 @@ VehicleDesc initVehicleDesc(PxMaterial* gMaterial, INIReader* config)
 	//Set up the chassis mass, dimensions, moment of inertia, and center of mass offset.
 	//The moment of inertia is just the moment of inertia of a cuboid but modified for easier steering.
 	//Center of mass offset is 0.65m above the base of the chassis and 0.25m towards the front.
-	const PxF32 chassisMass = config->GetReal("bike", "chassisMass", 1500.0f);
+	const PxF32 chassisMass = (float)config->GetReal("bike", "chassisMass", 1500.0);
 	const PxVec3 chassisDims(2.5f,2.0f,8.0f);
 	const PxVec3 chassisMOI
-		((chassisDims.y*chassisDims.y + chassisDims.z*chassisDims.z)*chassisMass/(config->GetReal("bike", "chassisMoIFactor", 12.0f)),
-		 (chassisDims.x*chassisDims.x + chassisDims.z*chassisDims.z)*0.8f*chassisMass/(config->GetReal("bike", "chassisMoIFactor", 12.0f)),
-		 (chassisDims.x*chassisDims.x + chassisDims.y*chassisDims.y)*chassisMass/(config->GetReal("bike", "chassisMoIFactor", 12.0f)));
+		((chassisDims.y*chassisDims.y + chassisDims.z*chassisDims.z)*chassisMass/(float)config->GetReal("bike", "chassisMoIFactor", 12.0),
+		 (chassisDims.x*chassisDims.x + chassisDims.z*chassisDims.z)*0.8f*chassisMass/(float)config->GetReal("bike", "chassisMoIFactor", 12.0),
+		 (chassisDims.x*chassisDims.x + chassisDims.y*chassisDims.y)*chassisMass/(float)config->GetReal("bike", "chassisMoIFactor", 12.0));
 
-	const PxF32 centerOfMassX = config->GetReal("bike", "centerOfMassX", 0.0);
-	const PxF32 centerOfMassY = config->GetReal("bike", "centerOfMassY", -chassisDims.y*0.5f - 0.65f);
-	const PxF32 centerOfMassZ = config->GetReal("bike", "centerOfMassZ", 0.0);
+	const PxF32 centerOfMassX = (float)config->GetReal("bike", "centerOfMassX", 0.0);
+	const PxF32 centerOfMassY = (float)config->GetReal("bike", "centerOfMassY", -chassisDims.y*0.5f - 0.65f);
+	const PxF32 centerOfMassZ = (float)config->GetReal("bike", "centerOfMassZ", 0.0);
 	const PxVec3 chassisCMOffset(centerOfMassX, centerOfMassY, centerOfMassZ);
 	//const PxVec3 chassisCMOffset(0.0f, -chassisDims.y*0.5f - 0.65f, 0.25f);
 
 	//Set up the wheel mass, radius, width, moment of inertia, and number of wheels.
 	//Moment of inertia is just the moment of inertia of a cylinder.
-	const PxF32 wheelMass = config->GetReal("bike", "wheelMass", 20.0);
-	const PxF32 wheelRadius = config->GetReal("bike", "wheelRadius", 0.5f);
-	const PxF32 wheelWidth = config->GetReal("bike", "wheelWidth", 0.4f);
-	const PxF32 wheelMOI = config->GetReal("bike", "wheelMoIFactor", 0.5f)*wheelMass*wheelRadius*wheelRadius;
+	const PxF32 wheelMass = (float)config->GetReal("bike", "wheelMass", 20.0);
+	const PxF32 wheelRadius = (float)config->GetReal("bike", "wheelRadius", 0.5f);
+	const PxF32 wheelWidth = (float)config->GetReal("bike", "wheelWidth", 0.4f);
+	const PxF32 wheelMOI = (float)config->GetReal("bike", "wheelMoIFactor", 0.5f)*wheelMass*wheelRadius*wheelRadius;
 	const PxU32 nbWheels = 4;
 
 	VehicleDesc vehicleDesc;
